Move feature specific code to modules
